Extract Direction enum from SnakeInterface to Directions.
