+++ SnakeController/Directions.hpp
@@ -1,4 +1,15 @@
+#pragma once
+
 namespace Snake {
+
+enum Direction
+{
+    Direction_UP    = 0b00,
+    Direction_DOWN  = 0b10,
+    Direction_LEFT  = 0b01,
+    Direction_RIGHT = 0b11
+};
+
 bool isHorizontal(Direction direction);
 bool isVertical(Direction direction);
 bool isPositive(Direction direction);
