+++ SnakeController/SnakeInterface.hpp
@@ -1,17 +1,10 @@
 #pragma once
+#include "Directions.hpp"
 
 #include <cstdint>
 
 namespace Snake
 {
-
-enum Direction
-{
-    Direction_UP    = 0b00,
-    Direction_DOWN  = 0b10,
-    Direction_LEFT  = 0b01,
-    Direction_RIGHT = 0b11
-};
 
 struct DirectionInd
 {
